move a few varfont checks out of googlefonts directory:
* fvar_name_entries
* mandatory_avar_table
* slant_direction
* varfont/consistent_axes
* varfont/duplexed_axis_reflow
* varfont/instances_in_order
* varfont/unsupported_axes
